Redo the packaging to package a slim package based on jpackage
This removes the maven-shade plugin and replaces it with the
tentackle-jlink-maven plugin. While it generates very slim and OS-specific
packages, one downside of jpackage is that it cannot cross-generate
packages between platforms.

However, if we tie it in with something like Travis or Github Actions
we can generate pacakges for all platforms while also being able to
control the Java runtime to bundle with the package (choosing for example
OpenJ9 will save us a lot of memory).

This also disables the Certificate code as one of the dependencies seem
to have problems with Jigsaw.