Properly implement the RPC service together with some model classes
This removes some of the old code and improves the way the RPC service
operates. For now, BlockCount (getblockcount), Info (getinfo) and
FrontendInfo (getfrontendinfo) are implemented and can be called via
the service. These classes describe both the Result and Request sections
of the API and each particular command.

This change also moves the loading of user credentials away from the
Daemon service and over to the RPC service as it makes slightly more
logical sense.