Fix the mess with monocle and use the classifier dependencies
This fixes the clash between the platform native dependencies and the
monocle ones that were occuring. Simply setting the classifier on a
platform (and set monocle) will include not only the monocle version, but
also the normal one - causing problems. Several exclusion rules were
needed to fix this.

This commit also moves the version and exclusion rules to the parent POM.