Make test suite compatible with the new model
Disabled the polling tests as they were triggering a bug in the polling
service. Whenever a new poll is registered, that class overwrites the
underlying timer. This means that we loose the handle to previous timers.

This also enables auto-discovery in CDI at the expense of some speed. It
was starting to get ridiculously verbose with the @WeldSetup annotation.